requirements.txt
streamlit
google-generativeai==0.8.5
pypdf
requests
beautifulsoup4
googlesearch-python
pandas
altair


